requirements.txt
python-dotenv
fastapi
uvicorn
psycopg2-binary